Calculating distance-squared when finding closest flight landing position now double precision (using ints isn't going to work for distances above about 50K.

+++ server/map.cpp
@@ -5402,11 +5402,11 @@ static char equal( GridPos inA, GridPos inB ) {
 
 
 static char tooClose( GridPos inA, GridPos inB, int inMinComponentDistance ) {
-    int xDist = inA.x - inB.x;
+    double xDist = (double)inA.x - (double)inB.x;
     if( xDist < 0 ) {
         xDist = -xDist;
         }
-    int yDist = inA.y - inB.y;
+    double yDist = (double)inA.y - (double)inB.y;
     if( yDist < 0 ) {
         yDist = -yDist;
         }
@@ -7024,9 +7024,9 @@ int addMetadata( int inObjectID, unsigned char *inBuffer ) {
 
 
 
-static unsigned int distSquared( GridPos inA, GridPos inB ) {
-    int xDiff = inA.x - inB.x;
-    int yDiff = inA.y - inB.y;
+static double distSquared( GridPos inA, GridPos inB ) {
+    double xDiff = (double)inA.x - (double)inB.x;
+    double yDiff = (double)inA.y - (double)inB.y;
     
     return xDiff * xDiff + yDiff * yDiff;
     }
@@ -7071,7 +7071,7 @@ GridPos getNextCloseLandingPos( GridPos inCurPose,
     
     int closestIndex = -1;
     GridPos closestPos;
-    unsigned int closestDist = INT_MAX;
+    double closestDist = DBL_MAX;
     
     for( int i=0; i<flightLandingPos.size(); i++ ) {
         GridPos thisPos = flightLandingPos.getElementDirect( i );
@@ -7084,7 +7084,7 @@ GridPos getNextCloseLandingPos( GridPos inCurPose,
 
         
         if( isInDir( inCurPose, thisPos, inDir ) ) {
-            unsigned int dist = distSquared( inCurPose, thisPos );
+            double dist = distSquared( inCurPose, thisPos );
             
             if( dist < closestDist ) {
                 // check if this is still a valid landing pos
@@ -7123,14 +7123,14 @@ GridPos getNextFlightLandingPos( int inCurrentX, int inCurrentY,
                                  doublePair inDir ) {
     int closestIndex = -1;
     GridPos closestPos;
-    unsigned int closestDist = INT_MAX;
+    double closestDist = DBL_MAX;
 
     GridPos curPos = { inCurrentX, inCurrentY };
 
     for( int i=0; i<flightLandingPos.size(); i++ ) {
         GridPos thisPos = flightLandingPos.getElementDirect( i );
         
-        unsigned int dist = distSquared( curPos, thisPos );
+        double dist = distSquared( curPos, thisPos );
         
         if( dist < closestDist ) {
